requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
bcrypt==4.2.0
black==24.3.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.2
deepdiff==7.0.1
distlib==0.3.8
dnspython==2.6.1
docker==7.1.0
ecdsa==0.19.0
email_validator==2.1.1
et-xmlfile==1.1.0
Faker==24.11.0
fastapi==0.110.2
filelock==3.13.1
freeze-requirements==0.5.3
freezegun==1.4.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
installer==0.7.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
loguru==0.7.2
MarkupSafe==2.1.5
mypy==1.9.0
mypy-extensions==1.0.0
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.10.7
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
phonenumbers==8.13.35
platformdirs==4.1.0
pluggy==1.5.0
psycopg2==2.9.9
pyasn1==0.6.0
pydantic==2.7.0
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pygame==2.5.2
PyJWT==2.9.0
pyparsing==3.1.1
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pywin32==306
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
ruff==0.1.14
sh==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.40.0
testcontainers==4.5.0
token-utils==0.1.8
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.29.0
virtualenv==20.25.0
watchfiles==0.24.0
websockets==13.0.1
win32-setctime==1.1.0
wrapt==1.16.0